fix initialization bugs in calls to LSQR and CGLS, update termination criteria in CGLS, fix bugs in calls to QR factorization in test_least_squares.m, fix construction of test5 for least squares, rename all files and folders to use lower case text

+++ utils/sketching_operators/sjlt.m
@@ -21,12 +21,12 @@ function[Omega] = sjlt(num_rows, num_cols, nnz)
     % nnz = 8;
     
     % Ensuring type compatabilities. 
-    num_cols = cast(num_cols, 'uint32');
-    num_rows = cast(num_rows, 'uint32');
-    nnz = cast(nnz, 'uint32');
+    num_cols = cast(num_cols, 'double');
+    num_rows = cast(num_rows, 'double');
+    nnz = cast(nnz, 'double');
     
     if num_cols >= num_rows
-        rows = uint32.empty;
+        rows = double.empty;
         nnz = min(num_cols, nnz);
         bad_size = num_rows < nnz;
         if bad_size
